(#17458) Re-enable performance fix
Previously, commit 20efe94 tried to only require puppet if it wasn't
loaded. However, defined?(Puppet) was always true when called by
absolute_path?, so the require was never executed.

Commit 327aff5, changed it to always require puppet, due to the way
puppet is required in 2.7.x. However, in 3.0.x, puppet is required at
the very beginning of command_line.rb (which is why 3.0.x didn't have
the same problem as 2.7.x).

This commit restores the behavior in 3.0.x, of never requiring puppet
from the absolute_path? method. I didn't want to make this change in the
merge commit, so it's clear why the logic is different.