(maint) Document Puppet::Parser::Functions.newfunction
Without this patch we have no YARD inline documentation for the
`Puppet::Parser::Functions.newfunction` class method.  This is a problem
because this is a very important user facing method that provides an API
that must remain stable across semantic versions.

This patch addresses the problem by documenting all of the options,
parameters, and user facing behaviors of the `newfunction` method.  In
addition, references to docs.puppetlabs.com, the stdlib module (which
heavily relies on this API), and the builtin template() function are
provided.

This patch also clearly documents the semantics of the new `:arity`
optional parameter.