(#17294) Remove empty puppet/provider/interface/base file
Without this patch there is a zero byte file located in the Puppet
library directory structure.  This is a problem because empty files are
considered errors when producing RPM packages.  I could not find any
reason this file exists, or any instances of this file being loaded.
Removing the file does not cause any apparent spec test failures.

This patch fixes the problem by removing the file from the tree.