(#17007) Make failonfail and combine explicit
Previously, many providers called Puppet::Util.execute with an
argument(s), which had the side-effect of defaulting failonfail and
combine to false.

This commit does not change any behavior. Instead in places where
execute is called with arguments, it explicitly sets failonfail and/or
combine to false. This way it's less confusing as to what the provider
is doing.