(#17007) Restore exec terminus behavior
Previously, the exec terminus attempts to call the execute method with
failonfail true semantics, but due to commit 8b76be30 for #3836, which
added the combine parameter, the execute method was actually being
called with failonfail false. As a result, the ExecutionFailure
exception was never being raised.

This commit restores the failonfail true semantics by making it
explicit, while preserving the combine => false behavior implemented
for #3836.