(#17007) Restore pre-3.0 command execution behavior
Prior to 3.0, commands were executed using Puppet::Util.execute
without any arguments, so failonfail and combine were true. In commit
89d57e72a, released in 3.0, provider commands were modified so that
the provider could supply options to the execute method. In doing so,
the default options were set to an empty hash, changing the default
failonfail and combine arguments to false.

This commit restores the pre-3.0 command execution behavior, such that
failonfail and combine are true.