(#11331) Add pops - this is the base implementation of new DSL model.
This adds a pops basic implementation of the DSL model, an evaluator,
the start of a validator, infrastructure classes and tests.

This commit contains the first working version of parsing. New test
parses a simple puppet DSL expressions and returns the expected result.

This adds proper handling of numbers:
- an lex error if a number is not formatted correctly
- parser creates QualifiedName or LiteralNumber for NAME token
- tree dumper outputs numbers with specified radix

Add support for match expr with quoted text as RHS.

This adds support for expressions like 'a'=~ "ends with ${foo}" to
interpolate into the regular expression (a string is compiled into a
regular expression).

Note that lexer now does not produce a REGEXP after INHERITS as it used
to in earlier versions. (3.1 flags inherits with regexp as syntax error
even if it is allowed in the grammar).

Add parse test for some 3.x "weird" class and define syntax.

This includes performing look ahead in the lexer (for {|) and emitting a
LAMBDA token instead of a LBRACE for the lambda start LBRACE.