(#17716) Add password managment on HPUX
This patch adds in basic trusted and non-trusted computing password
management for HPUX 11.31 IA. Code should work on 11.11 and all 11i
but has not been tested on those platforms specifically yet.  All
commands and files modified are present on 11.11 and 11i versions
and modified process is similar to past Perl scripts to do similar
actions on those aforementioned platforms.