(#7316) Use the public API for initializing settings
Previously the Puppet::Util::CommandLine code would use unpublished APIs
for initializing the application. Although this worked, it caused at
least two problems.
  1) The command line code was tied to the internals of puppet
  2) The public API could not actually be used for achieving what the
  command line code did

This commit changes the Puppet::Util::CommandLine class to start using
the public API of Puppet.initialize_settings for initializing puppet.
This required adding a new parameter to initialize_settings so that the
command line arguments could be passed in. It also required removing the
restriction that application_defaults can only be initialized once,
since the public behavior of the initialize_settings method is to
initialize both global and application settings at once.