(Maint) Create a mutation preserving wrapper for CommandLine
When the CommandLine class was changed to parse its options difference
and no longer hold onto a reference to the discovered options the
application option parsing code stopped working correctly because it
depended on being able to modify those values and have CommandLine hold
onto the modifications. This behavior caught be by surprise and took a
long time to track down where that occurred.

This commit preserves both the new behavior of CommandLine and the old
behavior of Application by creating a wrapper that can hold onto the
values from CommandLine in a manner that Application can modify, which
creates a separation between the two classes.