(#7316) Ensure changes to mutable CommandLine wrapper don't affect ARGV
In commit 1390faee94, we created a mutable CommandLine wrapper, but in
doing so, it had a reference to the original args that the CommandLine
was created with. However, several places in puppet and specs modify
`CommandLine#args`, such as `FaceBase#parse_options`. In particular, if
a CommandLine instance is created with the default set of arguments,
then FaceBase would mutate ARGV, causing order dependent test failures
that assumed ARGV was empty.

This commit ensures the CommandLine wrapper is constructed with a
duplicate set of arguments, that when modified, do not affect the
original set of arguments that the CommandLine was created with.

Paired-with: Andy Parker <andy@puppetlabs.com>