(#17554) Use agent run_mode when resolving runinterval
Previously, the windows service was invoking `puppet config print
runinterval` to determine how frequently to invoke the puppet agent.
However, that only takes into account the main section, not the agent
section, and runinterval is agent-specific.

This commit changes the windows service to invoke `puppet agent
--configprint runinterval` which takes into account the agent
section of the settings.