(#7316) Restore unknown subcommand behavior
The output of puppet with an unknown subcommand changed unexpectedly. If
version arguments were specified (-V or --version), then the puppet
version was no longer displayed. If no version arguments were specified,
then the "See 'puppet help' ..." message was no longer displayed.

This commit consolidates the version and help message behavior into a
NilSubcommand, and modifies the UnknownSubcommand to extend it, thereby
preserving the behavior that existed in 3.0.x.

It also adds tests for behavior that was only partially covered in the
acceptance test `ticket_17458_puppet_command_prints_help.rb`