Wrap user plist retrieval into get_users_plist()
Previously, the act of retrieving the user's plist was
performed on an ad-hoc basis by running the plutil command.
Because this binary is only available on OS X, it needed
to be stubbed out to successfully test its behavior on
Linux. To make this process easier (and also encapsulate
the behavior into its own method for future upgrades), the
plutil call has been wrapped in its own method
(get_users_plist).