(#6631) Remove runlevel 4 from default-start, add requires, buildrequires
As stated in the ticket, runlevel 4 is not defined on suse, so the init scripts
should not have a default-start on runlevel 4. Additionally, because we ship
logrotate rules in the package, the package needs to depend on logrotate. The
buildrequires of klogd and sysconfig are needed as buildtime dependencies as
noted in the ticket.