(#17845) - Add YARD doc to [] and []= in scope (= API)
The scope is passed to the Ruby DSL and is thus exposed to those that
author Ruby DSL logic. Since the same scope is also used by the Ruby DSL
implementation itself it is difficult to shield the user from access to
the real object. Instead, the methods that are API should be marked.

This commit adds "PUBLIC API" and a more doc text to the only two
methods that are considered to be API.

This required that the aliased methods [] and []= where changed to real
methods as it was not possible to mark the implemented methods as non
api while the aliases were api (there is a difference in parameter
passing to the operators and the underlying methods).