Warn when using --[no-]tests flags in install.rb
The tests flag isn't functional anymore, as there are no tests in the test
directory to run. This commit adds a warning for users of the tests flag that
the flag is useless. It also changes the default value for the tests flag to
false and promises that we will remove the flag entirely in the future.