(#17887) Only log if message has been initialized
Inside a test harness with Mocha, it's possible to run the "ensure" block
without having executed either the main or "rescue" blocks if an unexpected
method is executed etc.

If this happens during the property.sync, this causes an immediate exit from
the main block, but the "ensure" block is still run.  The event message won't
have been initialized yet, so this then causes the logger to report:

  Puppet::Util::Log requires a message

This changes the "ensure" block to only log if it's got as far as writing a
message.