Rescue ArgumentError in rbCFPropertyList.rb
Previously, rbCFPropertyList.rb had instantiated an object of
Enumerable::Enumerator and passed it an argument of an empty Array.
This caused issues when running Puppet in Ruby 1.9.3 because it threw
a stacktrace giving an ArgumentError (1 for 0) on the initialize
method of Enumerator.  There is an existing rescue block in the
rbCFPropertyList.rb file, but that wasn't getting caught because it
was only catching a NameError exception, and not an ArgumentError
exception.

This commit catches the ArgumentError exception class, which fixes
our problem.

Conflicts:
	lib/puppet/vendor/cfpropertylist/rbCFPropertyList.rb