(#18047) Revert change to user confdir
In 3.0, the default user confdir (and vardir) was changed from ~/.puppet
to the Dir::LOCAL_APPDATA/PuppetLabs. However, when puppet is run as a
non-administrator, puppet will attempt to create its confdir, e.g.
C:\Users\albert\AppData\Local\PuppetLabs\puppet, and fail because the
parent PuppetLabs directory doesn't exist.

This wasn't an issue in 2.7.x, because the confdir was ~/.puppet, so the
parent directory always existed.

This commit restores the previous behavior for non-administrators. No
changes are required for administrators, because the MSI installer creates
the necessary directories within COMMON_APPDATA.