(#17448/#17449/#18045) Update fileserver.conf and auth.conf examples
These files still referred to the regime whereby access control lived in
both files and auth.conf usually delegated to fileserver.conf, except
when it didn't. In the meantime, we've improved auth.conf and partially
broken fileserver.conf, and in conversations with the core team, it's
become clear that we don't necessarily WANT to fix fileserver.conf,
especially if not doing so will lead to better and more centralized
access control.

Unfortunately, an awkward period seems unavoidable. This commit
attempts to describe what users SHOULD do and give enough hints that it
will be easy to do so, at the cost of a certain windiness.

NOTE ESPECIALLY that we're now using the word "deprecated," as the
warnings-and-prefs-based deprecation plan outlined at #18045 has been accepted.