(#7316) Test faces from modules using a module
Previously the test that we could load faces from modules did so by
constructing a module in the configured modulepath. A more complete test
is to install a module that contains a face and ensure that we can
execute the face that has been installed in that manner. This commit
changes the test to do that.