(#14728) Handle removed files in the PMT changes command.
This patch ensures that "puppet module changes" command reports removed files
as changed rather than exiting with a "file not found" error message when it
encounters a file that should exist according to the metadata.json but which
is not present.
The patch also adds unit & acceptance tests for the "puppet module changes"
command.