16856 - Add a module data directory and hiera backend to consume it
The module HieraPuppet exist to be a helper that the various parts of
Puppet use to access Hiera - parts like the various parser functions and
the indirector.

Before this commit the indirector did not use this module instead it
duplicated a bunch of functionality and tests.

This commit refactors this to use the module and provide some
consistency and a single place to affect behavior changes