(#15496) Remove duplication of checks in ZAML tests
The tests previously were fairly disjoint and had many instances of
checking that a value round-tripped through YAML correctly. This starts
to pull them together and make use of common assertions. It also tries
using rspec matchers in order to achieve this. Ideally the matchers
would compose, but I don't see how to do that with the rspec interface.