(#18121) Collector overrides should work in Ruby 1.9
Increase arity of child_of? meta_def to 1

child_of? defined for Puppet::Resource::Type objects has an arity of 1,
but the meta_def of the same function created by
Puppet::Parser::Collector objects has an arity of 0. This throws errors
when the meta_def method is called with arguments. This commit matches
the arity so that this error will not be present.