(maint) Remove xargs rm on empty file list from puppet.spec.erb
There are no empty files in examples, so the `xargs rm` call on that directory
fail. This commit removes that line and adds a --no-run-if-empty to the other
xargs call in that section. This commit also moves conf/puppet-queue.conf to an
existing examples directory instead of examples/etc/puppet, which no longer
exists.