(yardoc) Add a @dsl tag to .yardopts.
This tag is intended to be used to mark methods that are part of an
internal DSL and which internal dsl they are part of. As an example;
internal DSL methods that deal with type definitions should be marked
with @dsl type.

In this commit the @dsl tag is not hidden in the output to make it
apparent where it has been set. The intent is to later hide this output.