(#14728) Declare stdout & stderr output expectations separately.
This patch updates the acceptance tests of the PMT changes command
such that the expectations on the stdout & stderr output are declared
separately. Originally when the expectations were declared together
the tests sometimes failed because the real output of the command
on the stdout and stderr was mixed in a slightly different order
than in the expectation.