(#17994) Add tests for directory creation
Adds tests to cover when creating directory and mode is specified as
either an octal or symbolic mode.

I was unable to actually create a directory (Dir.mkdir) and then stat it
to verify the mode was set correctly, because on some ruby
implementations, Dir.mkdir ignores special mode bits. On my Mac, for
example, calling Dir.mkdir(path, 01777), results in mode 00777. However,
puppet does the right thing, because it calls `property_fix` on the file
type, which goes through each property and ensures each is in sync.

Since the sticky-bit behavior is system-dependent, I'm setting an
expectation on the Dir object, and that at least detects the bug
reported in #17994.